+++ full-it/src/test/java/apoc/full/it/UUIDMultiDbTest.java
@@ -19,9 +19,7 @@
 package apoc.full.it;
 
 import apoc.util.Neo4jContainerExtension;
-import apoc.util.TestContainerUtil;
-import apoc.util.TestUtil;
-import org.junit.After;
+import apoc.util.TestContainerUtil.ApocPackage;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -41,15 +39,10 @@ import static apoc.ApocConfig.APOC_UUID_ENABLED;
 import static apoc.ApocConfig.APOC_UUID_ENABLED_DB;
 import static apoc.util.SystemDbTestUtil.TIMEOUT;
 import static apoc.util.TestContainerUtil.createEnterpriseDB;
-import static apoc.util.TestUtil.isRunningInCI;
 import static apoc.uuid.UUIDTestUtils.assertIsUUID;
-import static apoc.uuid.UuidHandler.APOC_UUID_REFRESH;
 import static apoc.uuid.UuidHandler.NOT_ENABLED_ERROR;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assume.assumeFalse;
-import static org.junit.Assume.assumeNotNull;
-import static org.junit.Assume.assumeTrue;
 import static org.neo4j.configuration.GraphDatabaseSettings.SYSTEM_DATABASE_NAME;
 import static org.neo4j.test.assertion.Assert.assertEventually;
 
@@ -57,28 +50,26 @@ public class UUIDMultiDbTest {
 
     private static Neo4jContainerExtension neo4jContainer;
     private static Driver driver;
-    private static final String dbTest = "dbtest";
-    private static final String dbEnabled = "dbenabled";
+    private static Session neo4jSession;
+    private static Session dbTestSession;
+    private static final String DB_TEST = "dbtest";
+    private static final String DB_ENABLED = "dbenabled";
     private static SessionConfig SYS_CONF;
 
     @BeforeClass
     public static void setupContainer() {
-        assumeFalse(isRunningInCI());
-        TestUtil.ignoreException(() -> {
-            neo4jContainer = createEnterpriseDB(List.of(TestContainerUtil.ApocPackage.FULL), true)
-                    .withEnv(Map.of(String.format(APOC_UUID_ENABLED_DB, dbTest), "false",
-                            APOC_UUID_ENABLED, "true",
-                            APOC_UUID_REFRESH, "1000"));
-            neo4jContainer.start();
-        }, Exception.class);
-        assumeNotNull(neo4jContainer);
-        assumeTrue("Neo4j Instance should be up-and-running", neo4jContainer.isRunning());
-
+        neo4jContainer = createEnterpriseDB(List.of(ApocPackage.FULL), true)
+                .withEnv(String.format(APOC_UUID_ENABLED_DB, DB_TEST), "false")
+                .withEnv(APOC_UUID_ENABLED, "true");
+        neo4jContainer.start();
         driver = neo4jContainer.getDriver();
+        createDatabases();
+        createSessions();
+
         SYS_CONF = SessionConfig.forDatabase(SYSTEM_DATABASE_NAME);
 
         try (Session session = driver.session()) {
-            Stream.of(dbTest, dbEnabled).forEach(
+            Stream.of(DB_TEST, DB_ENABLED).forEach(
                     db -> session.run(String.format("CREATE DATABASE %s", db))
             );
         }
@@ -86,79 +77,53 @@ public class UUIDMultiDbTest {
 
     @AfterClass
     public static void bringDownContainer() {
-        if (neo4jContainer != null && neo4jContainer.isRunning()) {
-            neo4jContainer.close();
-        }
-    }
-
-    @After
-    public void after() {
-        try (Session session = driver.session(SYS_CONF)) {
-            Stream.of("neo4j", dbEnabled).forEach(
-                    db -> session.run("CALL apoc.uuid.dropAll($db)",
-                            Map.of("db", db))
-            );
-        }
+        neo4jContainer.close();
     }
 
     @Test(expected = RuntimeException.class)
-    public void testWithSpecificDatabaseWithUUIDDisabled() throws Exception {
-
-        Session session = driver.session(SessionConfig.forDatabase(dbTest));
+    public void testWithSpecificDatabaseWithUUIDDisabled() {
         try {
-
-            session.writeTransaction(tx -> tx.run(
-                    "CREATE CONSTRAINT ON (foo:Foo) ASSERT foo.uuid IS UNIQUE")
-            );
-
-            session.writeTransaction(tx -> tx.run(
-                    "CALL apoc.uuid.install('Foo') YIELD label RETURN label")
-            );
-
+            dbTestSession.writeTransaction(tx -> tx.run("CREATE CONSTRAINT FOR (foo:Foo) REQUIRE foo.uuid IS UNIQUE"));
+            dbTestSession.writeTransaction(tx -> {
+                tx.run("CREATE (d:Foo {name:'Test'})-[:WORK]->(l:Bar {name:'Baz'})");
+                return tx.run("CALL apoc.uuid.install('Foo', {addToExistingNodes: false }) YIELD label RETURN label");
+            });
         } catch (RuntimeException e) {
             String expectedMessage = "Failed to invoke procedure `apoc.uuid.install`: " +
-                    "Caused by: java.lang.RuntimeException: " + String.format(NOT_ENABLED_ERROR, dbTest);
+                                     "Caused by: java.lang.RuntimeException: " + String.format(NOT_ENABLED_ERROR, DB_TEST);
             assertEquals(expectedMessage, e.getMessage());
             throw e;
         }
     }
 
     @Test
-    public void testWithDefaultDatabaseWithUUIDEnabled() throws InterruptedException {
-        try (Session session = driver.session(SessionConfig.forDatabase("neo4j"))) {
+    public void testWithDefaultDatabaseWithUUIDEnabled() {
+        neo4jSession.writeTransaction(tx -> tx.run("CREATE CONSTRAINT FOR (foo:Foo) REQUIRE foo.uuid IS UNIQUE"));
+        neo4jSession.writeTransaction(tx -> {
+            tx.run("CALL apoc.uuid.install('Foo', {addToExistingNodes: false }) YIELD label RETURN label");
+            return tx.run("CREATE (d:Foo {name:'Test'})-[:WORK]->(l:Bar {name:'Baz'})");
+        });
 
-            session.writeTransaction(tx -> tx.run(
-                    "CREATE CONSTRAINT ON (foo:Foo) ASSERT foo.uuid IS UNIQUE")
-            );
+        String call = "MATCH (n:Foo) RETURN n.uuid as uuid";
+        AtomicBoolean nodeHasUUID = new AtomicBoolean(false);
+        Consumer<Iterator<Map<String, Object>>> resultConsumer = (result) -> {
+            Map<String, Object> r = result.next();
+            nodeHasUUID.set(r.get("uuid") != null);
+        };
 
-            session.writeTransaction(tx -> tx.run(
-                    "CALL apoc.uuid.install('Foo') YIELD label RETURN label")
-            );
-            session.writeTransaction(tx -> tx.run(
-                    "CREATE (d:Foo {name:'Test'})-[:WORK]->(l:Bar {name:'Baz'})")
-            );
-
-            String call = "MATCH (n:Foo) RETURN n.uuid as uuid";
-            AtomicBoolean nodeHasUUID = new AtomicBoolean(false);
-            Consumer<Iterator<Map<String, Object>>> resultConsumer = (result) -> {
-                Map<String, Object> r = result.next();
-                nodeHasUUID.set(r.get("uuid") != null);
-            };
-
-            long timeout = System.currentTimeMillis() + TimeUnit.SECONDS.toMillis(5);
-            while (System.currentTimeMillis() < timeout && !nodeHasUUID.get()) {
-                session.writeTransaction(tx -> {
-                    Map<String, Object> p = Collections.<String, Object>emptyMap();
-                    resultConsumer.accept(tx.run(call, p).list()
-                            .stream()
-                            .map(Record::asMap)
-                            .collect(Collectors.toList()).iterator());
-                    tx.commit();
-                    return null;
-                });
-            }
-            assertTrue("UUID not set on node after 5 seconds", nodeHasUUID.get());
+        long timeout = System.currentTimeMillis() + TimeUnit.SECONDS.toMillis(5);
+        while (System.currentTimeMillis() < timeout && !nodeHasUUID.get()) {
+            neo4jSession.writeTransaction(tx -> {
+                Map<String, Object> p = Collections.<String, Object>emptyMap();
+                resultConsumer.accept(tx.run(call, p).list()
+                        .stream()
+                        .map(org.neo4j.driver.Record::asMap)
+                        .collect(Collectors.toList()).iterator());
+                tx.commit();
+                return null;
+            });
         }
+        assertTrue("UUID not set on node after 5 seconds", nodeHasUUID.get());
     }
 
     //
@@ -169,11 +134,11 @@ public class UUIDMultiDbTest {
     public void createUUIDWithSpecificDbWithUUIDDisabled() {
         try(Session session = driver.session(SYS_CONF)) {
             session.run("CALL apoc.uuid.setup('Baz', $db) YIELD label RETURN label",
-                    Map.of("db", dbTest)
+                    Map.of("db", DB_TEST)
             );
         } catch (RuntimeException e) {
             String expectedMessage = "Failed to invoke procedure `apoc.uuid.setup`: " +
-                    "Caused by: java.lang.RuntimeException: " + String.format(NOT_ENABLED_ERROR, dbTest);
+                    "Caused by: java.lang.RuntimeException: " + String.format(NOT_ENABLED_ERROR, DB_TEST);
             assertEquals(expectedMessage, e.getMessage());
             throw e;
         }
@@ -204,9 +169,9 @@ public class UUIDMultiDbTest {
     public void createUUIDInNotDefaultDbWithUUIDEnabled() throws InterruptedException {
         driver.session(SYS_CONF)
                 .run("CALL apoc.uuid.setup('Another', $db)",
-                        Map.of("db", dbEnabled));
+                        Map.of("db", DB_ENABLED));
 
-        try(Session session = driver.session(SessionConfig.forDatabase(dbEnabled))) {
+        try(Session session = driver.session(SessionConfig.forDatabase(DB_ENABLED))) {
             // check uuid set
             awaitUuidSet(session, "Another");
 
@@ -238,4 +203,14 @@ public class UUIDMultiDbTest {
         }, (val) -> val, TIMEOUT, TimeUnit.SECONDS);
     }
 
+    private static void createDatabases() {
+        try (Session systemSession = driver.session(SessionConfig.forDatabase("system"))) {
+            systemSession.writeTransaction(tx -> tx.run(String.format("CREATE DATABASE %s WAIT;", DB_TEST)));
+        }
+    }
+
+    private static void createSessions() {
+        neo4jSession = neo4jContainer.getSession();
+        dbTestSession = driver.session(SessionConfig.forDatabase(DB_TEST));
+    }
 }
