+++ test-utils/src/main/java/apoc/util/GoogleCloudStorageContainerExtension.java
@@ -21,11 +21,6 @@ package apoc.util;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
-import org.testcontainers.containers.wait.strategy.LogMessageWaitStrategy;
-import org.testcontainers.containers.wait.strategy.WaitAllStrategy;
-import org.testcontainers.containers.wait.strategy.WaitStrategy;
-
-import java.time.Duration;
 
 import static java.net.HttpURLConnection.HTTP_OK;
